requirements.txt
pandas
pyyaml
mysql-python
mecab-python
slackclient
